% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_BF.R
\name{BF_compute}
\alias{BF_compute}
\title{Compute Bayes Factors for Each Dataset in a FASH Object}
\usage{
BF_compute(fash)
}
\arguments{
\item{fash}{A \code{fash} object containing the fitted model and likelihood matrix.}
}
\value{
A numeric vector of Bayes Factors, where each entry corresponds to a dataset.
}
\description{
This function computes Bayes Factors (BF) for each dataset in a \code{fash} object.
The BF is calculated as the ratio of likelihood under the alternative hypothesis
to the likelihood under the null hypothesis.
}
\examples{

data_list <- list(
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0),
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0)
)
grid <- seq(0, 2, length.out = 10)
fash_obj <- fash(data_list = data_list, Y = "y", smooth_var = "x", grid = grid, likelihood = "poisson", verbose = TRUE)

# Compute Bayes Factors
BF_values <- BF_compute(fash_obj)
print(BF_values)

}
