% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_BF.R
\name{BF_control}
\alias{BF_control}
\title{Perform Bayes Factor-Based Control for Estimating \eqn{\pi_0}}
\usage{
BF_control(BF, plot = FALSE)
}
\arguments{
\item{BF}{A numeric vector of Bayes Factors computed from `BF_compute()`.}

\item{plot}{A logical value. If `TRUE`, generates diagnostic plots for BF control. Default is `FALSE`.}
}
\value{
A list containing:
\describe{
  \item{mu}{Cumulative mean of sorted Bayes Factors.}
  \item{pi0_hat}{Estimated \eqn{\pi_0} values for each BF threshold.}
  \item{pi0_hat_star}{Final estimated \eqn{\pi_0} based on the first BF threshold where \eqn{E(BF) \geq 1}.}
}
}
\description{
This function estimates **\eqn{\pi_0}**, the proportion of datasets that follow the null hypothesis,
using **Bayes Factor (BF) control**.
}
\examples{
BF_values <- runif(100, 0.5, 5)  # Example Bayes Factors
BF_control_results <- BF_control(BF_values, plot = TRUE)
print(BF_control_results$pi0_hat_star)

}
