% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_likelihood.R
\name{compute_L_poisson_helper}
\alias{compute_L_poisson_helper}
\title{Compute Log-Likelihood for a Poisson TMB Model}
\usage{
compute_L_poisson_helper(
  data_i,
  psd_iwp,
  num_basis = 30,
  betaprec = 1e-06,
  order = 2,
  pred_step = 1
)
}
\arguments{
\item{data_i}{A single dataset extracted from the \code{data_list} component of \code{fash_set_data}.
Must be a list containing \code{y}, \code{x}, and \code{offset}.}

\item{psd_iwp}{A numeric value for the precision parameter of the Integrated Wiener Process. If 0, only fixed effects are used.}

\item{num_basis}{An integer specifying the number of O-Spline basis functions.}

\item{betaprec}{A numeric value representing the precision of the fixed effects coefficients (`beta`).}

\item{order}{An integer specifying the order of the Integrated Wiener Process (IWP) prior.}

\item{pred_step}{A numeric value specifying the prediction step size.}
}
\value{
A numeric value representing the negative log-likelihood for the dataset.
}
\description{
Computes the log-likelihood for a Poisson model using a TMB-based approach.
Handles both random-effects models and fixed-effects-only models.
}
\examples{
# Example usage
Y <- matrix(rpois(20, lambda = 5), nrow = 4, ncol = 5)
smooth_var <- matrix(runif(20), nrow = 4, ncol = 5)
offset <- 1
fash_data <- fash_set_data(Y, smooth_var, offset)
log_likelihood <- compute_L_poisson_helper(
  data_i = fash_data$data_list[[1]], psd_iwp = 0.1
)

}
\keyword{internal}
