% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_lfsr.R
\name{compute_marginal_mean_var_once}
\alias{compute_marginal_mean_var_once}
\title{Compute Marginal Mean and Variance for a Single PSD Value}
\usage{
compute_marginal_mean_var_once(
  data_i,
  refined_x,
  psd_iwp,
  Si = NULL,
  Omegai = NULL,
  num_basis = 30,
  betaprec = 1e-06,
  order = 2,
  pred_step = 1,
  likelihood,
  deriv = 0
)
}
\description{
Computes the posterior mean and variance for a given dataset, refined x-values,
and a specific PSD value.
}
\keyword{internal}
