% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_lfsr.R
\name{compute_posterior_sign_prob}
\alias{compute_posterior_sign_prob}
\title{Compute Probability of Being Positive or Negative}
\usage{
compute_posterior_sign_prob(mu, sigma2)
}
\description{
Given posterior means and variances, this function computes the probability
that the posterior distribution is positive or negative at each point.
}
\keyword{internal}
