% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_likelihood.R
\name{fash_L_compute}
\alias{fash_L_compute}
\title{Compute the L Matrix for a Set of Datasets}
\usage{
fash_L_compute(
  fash_data,
  likelihood = "gaussian",
  num_cores = 1,
  grid = seq(0, 2, length.out = 10),
  pred_step = 1,
  num_basis = 30,
  betaprec = 1e-06,
  order = 2,
  verbose = FALSE
)
}
\arguments{
\item{fash_data}{The output from `fash_set_data`, containing preprocessed datasets.}

\item{likelihood}{A character string specifying the likelihood function to use. Options are `"gaussian"` and `"poisson"`. Default is `"gaussian"`.}

\item{num_cores}{An integer specifying the number of cores to use for parallel processing. Default is 1.}

\item{grid}{A numeric vector representing the grid of PSD (Predictive Standard Deviation) values. Default is `seq(0, 2, length.out = 10)`.}

\item{pred_step}{A numeric value specifying the prediction step size. Default is 1.}

\item{num_basis}{An integer specifying the number of O-Spline basis functions to use for the approximation. Default is 30.}

\item{betaprec}{A numeric value representing the precision of the fixed effects coefficients (`beta`). Default is `1e-6`.}

\item{order}{An integer specifying the order of the Integrated Wiener Process (IWP) prior. Default is 2.}

\item{verbose}{A logical value. If `TRUE`, shows a progress bar when `num_cores = 1`. Default is `FALSE`.}
}
\value{
A numeric matrix where each row corresponds to a dataset and each column corresponds to a grid value of PSD.
}
\description{
Computes the L matrix, where each row corresponds to a dataset and each column corresponds to a grid value of PSD (Predictive Standard Deviation).
Handles both Gaussian and Poisson likelihoods using helper functions.
}
\examples{
# Example usage
Y <- matrix(rpois(20, lambda = 5), nrow = 4, ncol = 5)
smooth_var <- matrix(runif(20), nrow = 4, ncol = 5)
offset <- 1
fash_data <- fash_set_data(Y, smooth_var, offset)
grid <- seq(0.1, 2, length.out = 10)
L_matrix_gaussian <- fash_L_compute(fash_data, likelihood = "gaussian", num_cores = 1, grid = grid, num_basis = 30, betaprec = 1e-6, order = 2, verbose = TRUE)
L_matrix_poisson <- fash_L_compute(fash_data, likelihood = "poisson", num_cores = 1, grid = grid, num_basis = 30, betaprec = 1e-6, order = 2, verbose = TRUE)

}
