% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fash.R
\name{fdr_control}
\alias{fdr_control}
\title{Perform False Discovery Rate (FDR) Control}
\usage{
fdr_control(fash_obj, alpha = 0.05, plot = FALSE)
}
\arguments{
\item{fash_obj}{A \code{fash} object containing the results of the FASH pipeline, including \code{lfdr}.}

\item{alpha}{A numeric value specifying the significance level for hypothesis testing.}

\item{plot}{A logical value. If \code{TRUE}, plots the sorted FDR values with a horizontal line at the \code{alpha} level.}
}
\value{
A list containing:

      - \code{fdr_results}: A data frame with columns:

      - \code{index}: The original dataset index.

      - \code{FDR}: The false discovery rate for each dataset.
}
\description{
This function performs hypothesis testing by controlling the False Discovery Rate (FDR) based on the
local false discovery rate (lfdr) stored in the \code{fash} object.
}
\examples{
# Example usage
set.seed(1)
data_list <- list(
  data.frame(y = rnorm(n = 5, sd = 0.5), x = 1:5, offset = 0),
  data.frame(y = rnorm(n = 5, sd = 0.8), x = 1:5, offset = 0),
  data.frame(y = rnorm(n = 5, sd = 0.6), x = 1:5, offset = 0),
  data.frame(y = rnorm(n = 5, sd = 0.7), x = 1:5, offset = 0)
)
S <- list(rep(0.5, 5), rep(0.8, 5), rep(0.6, 5), rep(0.7, 5))
grid <- seq(0, 2, length.out = 10)
fash_obj <- fash(data_list = data_list, Y = "y", smooth_var = "x", offset = "offset", S = S, grid = grid, likelihood = "gaussian", verbose = TRUE)
fdr_control(fash_obj, alpha = 0.05, plot = TRUE)

}
