% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fashr.R
\name{plot.fashr}
\alias{plot.fashr}
\title{Plot Method for FASHR Objects}
\usage{
\method{plot}{fashr}(x, ordering = NULL, discrete = FALSE, ...)
}
\arguments{
\item{x}{A `fashr` object containing the results of the FASH pipeline.}

\item{ordering}{A character string specifying the method for reordering datasets (e.g., "mean" or "lfdr").

  - `"mean"`: Reorder by the mean of the posterior PSD.

  - `"lfdr"`: Reorder by the local false discovery rate.

  - `NULL`: No reordering (default).}

\item{discrete}{A logical value. If `TRUE`, treats PSD values as discrete categories with distinct colors.
If `FALSE`, treats PSD values as a continuous variable with a gradient. Default is `FALSE`.}

\item{...}{Additional arguments passed to `fash_structure_plot`.}
}
\description{
Generates a structure plot of the posterior weights stored in the `fashr` object,
visualizing the distribution of posterior weights across datasets and PSD values.
}
\examples{
data_list <- list(
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0),
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0)
)
S <- NULL
Omega <- NULL
grid <- seq(0, 2, length.out = 10)
fashr_obj <- fashr(data_list = data_list, Y = "y", smooth_var = "x", offset = "offset", S = S, Omega = Omega, grid = grid, likelihood = "poisson", verbose = TRUE)
plot(fashr_obj, ordering = "mean", discrete = TRUE)

}
