% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fash.R
\name{print.fash}
\alias{print.fash}
\title{Print Method for fash Objects}
\usage{
\method{print}{fash}(x, ...)
}
\arguments{
\item{x}{A \code{fash} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Displays a summary of the fitted \code{fash} object, including the number of datasets,
type of likelihood used, the number of PSD grid values, and the order of the Integrated Wiener Process (IWP).
}
\examples{
set.seed(1)
data_list <- list(
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0),
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0)
)
grid <- seq(0, 2, length.out = 10)
fash_obj <- fash(data_list = data_list, Y = "y", smooth_var = "x", offset = "offset", grid = grid, likelihood = "poisson", verbose = TRUE)
print(fash_obj)

}
