% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fashr.R
\name{print.fashr}
\alias{print.fashr}
\title{Print Method for FASHR Objects}
\usage{
\method{print}{fashr}(x, ...)
}
\arguments{
\item{x}{A `fashr` object.}

\item{...}{Additional arguments (not used).}
}
\description{
Displays a summary of the fitted `fashr` object, including the number of datasets,
type of likelihood used, the number of PSD grid values, and the order of the Integrated Wiener Process (IWP).
}
\examples{
data_list <- list(
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0),
  data.frame(y = rpois(5, lambda = 5), x = 1:5, offset = 0)
)
grid <- seq(0, 2, length.out = 10)
fashr_obj <- fashr(data_list = data_list, Y = "y", smooth_var = "x", offset = "offset", grid = grid, likelihood = "poisson", verbose = TRUE)
print(fashr_obj)

}
