% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_simulate.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Observed Data from a Specified True Function}
\usage{
simulate_data(g, x = NULL, sd = 0.1)
}
\arguments{
\item{g}{A function specifying the true underlying signal to simulate.}

\item{x}{A numeric vector of grid points where the function \code{g} is evaluated.}

\item{sd}{A numeric value or vector specifying the standard deviation(s) of noise.
If a scalar, all points have the same standard deviation. If a vector, it is
sampled with replacement for each \code{x}.}
}
\value{
A data frame containing:
  \describe{
    \item{\code{x}}{Input grid points.}
    \item{\code{y}}{Simulated noisy observations.}
    \item{\code{truef}}{True function values at \code{x}.}
    \item{\code{sd}}{Standard deviation used for noise at each point.}
  }
}
\description{
This function generates synthetic observed data by evaluating a user-specified
true function \code{g} on a grid of x-values and adding Gaussian noise
with heteroscedastic standard deviation sampled from a provided vector or scalar.
}
\examples{
g <- function(x) sin(x / 5)
fashr:::simulate_data(g, sd = 0.2)

}
\keyword{internal}
