% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_simulate.R
\name{simulate_linear_function}
\alias{simulate_linear_function}
\title{Simulate a Random Linear Function}
\usage{
simulate_linear_function(sd_poly = 1, pred_step = 16)
}
\arguments{
\item{sd_poly}{Standard deviation for both intercept and slope terms.}

\item{pred_step}{Resolution scaling for the slope variance.}
}
\value{
A function that evaluates the random linear function at new x-values.
}
\description{
Generates a random linear function with intercept and slope drawn from normal distributions.
}
\examples{

f <- fashr:::simulate_linear_function()
plot(1:16, f(1:16), type = 'l')

}
\keyword{internal}
