% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_simulate.R
\name{simulate_nondynamic_function}
\alias{simulate_nondynamic_function}
\title{Simulate a Random Constant Function (Nondynamic)}
\usage{
simulate_nondynamic_function(sd_poly = 1)
}
\arguments{
\item{sd_poly}{Standard deviation for the constant term.}
}
\value{
A function that evaluates to a constant at any x.
}
\description{
Generates a constant function with value drawn from a normal distribution.
}
\examples{

f <- fashr:::simulate_nondynamic_function()
plot(1:16, f(1:16), type = 'l')

}
\keyword{internal}
