% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_simulate.R
\name{simulate_nonlinear_function}
\alias{simulate_nonlinear_function}
\title{Simulate a Random Nonlinear Function from an Integrated Wiener Process (IWP) Prior}
\usage{
simulate_nonlinear_function(
  n_basis = 20,
  sd_function = 1,
  sd_poly = 0.1,
  p = 1,
  pred_step = 16,
  x_range = NULL
)
}
\arguments{
\item{n_basis}{An integer specifying the number of spline basis functions (minimum 3).}

\item{sd_function}{A numeric value representing the predictive standard deviation of the function.}

\item{sd_poly}{A numeric value specifying the standard deviation for the linear polynomial component.}

\item{p}{An integer specifying the order of the polynomial trend component (default is 1 for linear).}

\item{pred_step}{A numeric value defining the prediction step size that determines the predictive standard deviation.}

\item{x_range}{A numeric vector of length 2 specifying the range of x-values over which the function is defined.}
}
\value{
A function that can evaluate the generated random smooth signal at any new \code{x} values:
  \describe{
    \item{function}{Returns the evaluated values of the sampled nonlinear function at specified \code{x} inputs.}
  }
}
\description{
This function generates a random smooth function sampled from an Integrated Wiener Process (IWP) prior,
parameterized by the number of basis functions and smoothing parameters.
}
\examples{
f <- fashr:::simulate_nonlinear_function()
plot(1:16, f(1:16), type = 'l')

}
\keyword{internal}
