% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_simulate.R
\name{simulate_quadratic_function}
\alias{simulate_quadratic_function}
\title{Simulate a Random Quadratic Function}
\usage{
simulate_quadratic_function(sd_poly = 1)
}
\arguments{
\item{sd_poly}{Standard deviation for the coefficients.}
}
\value{
A function that evaluates the random quadratic function at new x-values.
}
\description{
Generates a random quadratic function with coefficients drawn from normal distributions.
}
\examples{

f <- fashr:::simulate_quadratic_function()
plot(1:16, f(1:16), type = 'l')

}
\keyword{internal}
